Using saveDefaultConfig function
There is an error with the plugin, every time you restart your server the plugin tries to re-add the config. When there is already a config file there. This calls a function to write the config, but only if there isn't a config there to begin with. This gets rid of that warning when you start the server saying there is already a config file there.